Change presence state API to take a state information when indicating changes.
Much like the improved ast_devstate_changed_* functions for device states,
it is much simpler if we provide a state, subtype, and message directly
when presence state changes. This way, the provider does not have to be
asked what changes have been made since in 99% of cases, the provider is
the one who is actually the one telling of the state change in the first
place.