Handle simple configs with very little code
This definitely needs some cleanup, but I just wanted to save it before
tearing into it again.

Basically, this allows one to set a bunch of alloc callbacks and some
other info, then load a config that has no relations between global
(e.g. "general") and specific (e.g. "peers") and handle config loading
and reloading with a single function call.

I want to find a way to clean up having the void * return types in
some of the callbacks since that means sillyness on the module side
of things. I could just make things void * and do casting on the other
side, but I hate doing that. I could cast at the module side, but that
is ugly too. We'll see.

Right now I have hardcoded "general" as the category with global options,
it will be a regex when I'm done.

I may decide I hate this.