Saving work.
Things work, there dont' appear to be any ref/memory leaks.
MUCH cleanup work to do this weekend. Then on to reviewboard.

This basically adds the ability to support multiple global
or private containers for configs--which may not be useful
to the majority of configs.

I'd also like to implement
1) matching config option objects by fields (to handle things
   like type=peer and type=user being different objects
2) forcing preloading of certain contexts so that users can
   make decisions after certain contexts are loaded. This
   would allow supporting legacy configs that inherit private
   options from