Fix test failure and add some cleanup to presence change test.
Switching to use an enum instead of an int for presence state values
made tests for results being less than 0 meaningless since enums are
treated as unsigned values. The fix for this is to create an enum value
for an invalid presence. Comparison to this value tells if the presence
is valid or bogus.

In addition, the presence change unit test has been changed to use a more
"unique" value for setting presence and then cleaning the DB afterwards.