Remove AST_CONTROL_CUSTOM frame type and add a public function to send a SIP INFO request.
In 1.8-digiumphones and 10-digiumphones, an AST_CONTROL_CUSTOM frame type was created in order
to send a channel driver-specific indication. In addition, functions were added in order to
encode SIP INFO data into a frame and decode data from a frame.

The code in question had an odd mix of generic behavior and very SIP-specific behavior. Since this
was added with the primary purpose of being able to send SIP INFO requests to a phone, it seemed
like the generic elements should be eliminated in preference of a more direct method of sending
a SIP INFO.

With this change, a new file, sip_api.h has been added with a single function, ast_sipinfo_send().
The function is defined in chan_sip.c. This means that any modules that use ast_sipinfo_send() will
need to load after chan_sip.so has been loaded.

Several files have been removed since they are no longer useful.