Get this compiling so bamboo will shut up.
At this point, adding message IDs to messages stored on the file
system should work under nominal conditions. There is some error handling
I need to add.

The difficult part is going to be getting this to work properly for
ODBC and IMAP. First, there is no current method for updating message data.
One can currently only store, rename, or delete. I will use the RENAME macro
as a guideline for updating message data since that is similar. Second, and
more importantly, it appears that ODBC and IMAP were never updated to actually
store or retrieve message IDs in the first place, so that needs to be added
into their STORE and RETRIEVE implementations.