Some simplification work
This is just a checkin of work and nowhere near actual readyness. Just a backup.
Still need to clean up the ao2 global stuff, going to make aco_type a non-ao2 object
that is just stored in a list or array in the aco_info. Mostly I wanted to commit
this because even though it isn't clean, it still functions and I might want to
revert back to it after messing stuff up next. :-p