Death to boiler plate code
Removes a lot of the boiler-plate allocation stuff for aco_types by having
them just be stack allocated--they never really change and exist for the
lifetime of the module.

Since I'm already using a variable-length array in aco_info for the "preload"
field (which won't be used as often as types), I can't easily define the types
field without an intermediary step. I'll probably switch the variable length
array to handling the types, then come up with something different for preload.