chan_sip: Check the right channel's host address for directmediapermit/deny
Prior to this patch, when checking the addresses for directmediapermit and
denydirectmediadeny, Asterisk would check the host address of the channel
permit/deny was specified, which defers from the expectations of both
our users and the development team. Instead, directmediapermit/deny now
checks against the address of the channel that the peer with the ACL is
connected to.

(issue AST-876)
Review: https://reviewboard.asterisk.org/r/1899/
........

Merged revisions 366547 from http://svn.asterisk.org/svn/asterisk/branches/1.8