OMGOMGOMG this is soo much better
No more shall there be separate private and private config containers.

Just because you can't store the config data on the "private" object
(because a failure in parsing the config would mean that you've updated
some of the privates, but not others after a reload), it doesn't mean
that you can't store a pointer to the private on the config data (which
is recreated from scratch after every reload, and therefor not a problem).

Basically, instead of separating out the config from the private state,
I now realize that the "private" (which I'm totally renaming "item" later
because private is a horrible name that confuses people because it means
something else in Asterisk) *is* the config data, and that any state
that must exist after reload is the object that must be "separated".

Since the state object is an refcounted object, when allocing a private
config object (item!), if there is an existing object, we can just keep
the old state around, reffing it. The code ends up being much simpler.

</ramble>