Working on negotating the various options here.
If the other side REQUIREs PRACK and we don't support it, we should answer negatively.
If the other side SUPPORT PRACK and we support it, we should activate it for this dialog.
If the other side do not support PRACK and we support it, do not activate it for this dialog.
If the other side SUPPORT PRACK and we don't, don't activate either