Naive configuration implementation.
This hasn't been too well thought out at this point,
but this now uses the configuration framework to set
some basic options.

Next thing I'm going to do is to address the configuration
changes in agent login. The agent structure contains state
information, so there is no need for it to exist outside of
a call to agent_login. Therefore, the state machine will likely
undergo a change to get rid of the "logged out" state since
the agent will not actually exist in that state. The current
logic that exists for entering and exiting the logged out state
will instead happen during agent allocation and destruction.